Fixed QMUN-1895

+++ Q-municate_core/src/main/java/com/quickblox/q_municate_core/utils/ChatUtils.java
@@ -163,7 +163,7 @@ public class ChatUtils {
     public static QMUser getOpponentFromPrivateDialog(QMUser currentUser, List<DialogOccupant> occupantsList) {
         for (DialogOccupant dialogOccupant : occupantsList) {
             if (dialogOccupant != null && dialogOccupant.getUser() != null
-                    && currentUser.getId() != dialogOccupant.getUser().getId()) {
+                    && currentUser.getId().intValue() != dialogOccupant.getUser().getId().intValue()) {
                 return dialogOccupant.getUser();
             }
         }
